Show the target URL in the replay info
