+++ src/ui/components/Library/PendingTeamScreen.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { PendingWorkspaceInvitation } from "ui/types";
+import { getDisplayedUrl } from "ui/utils/environment";
 import PendingTeamPrompt from "./PendingTeamPrompt";
-import { getDisplayedUrl, getDurationString, getRelativeDate } from "./RecordingRow";
+import { getDurationString, getRelativeDate } from "./RecordingRow";
 
 const MOCK_DATA = [
   { date: "2021-12-01T18:37:44.077Z", user: { name: "Jaril" } },
